Dto/Entity и всякое

+++ src/main/java/org/example/services/PersonService.java
@@ -1,7 +1,16 @@
 package org.example.services;
 
+import jakarta.persistence.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.example.models.Book;
-import org.example.models.Person;
+
+import org.example.models.dto.BookDto;
+import org.example.models.dto.PersonDto;
+import org.example.models.entity.BookEntity;
+import org.example.models.entity.PersonEntity;
+import org.example.models.mapper.BookMapper;
+import org.example.models.mapper.PersonMapper;
 import org.example.repositories.BookRepository;
 import org.example.repositories.PersonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,50 +19,62 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.time.Period;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 
+@Slf4j
 @Service
 @Transactional(readOnly = true)
-public class PersonService{
+@RequiredArgsConstructor
+public class PersonService {
 
     private final PersonRepository personRepository;
     private final BookRepository bookRepository;
+    private final PersonMapper personMapper;
+    private final BookMapper bookMapper;
 
-    @Autowired
-    public PersonService(PersonRepository personRepository, BookRepository bookRepository) {
-        this.personRepository = personRepository;
-        this.bookRepository = bookRepository;
+    public List<PersonDto> findAll() {
+        List<PersonEntity> people = personRepository.findAll();
+        return people.stream().map(personMapper::toDto).toList();
     }
 
-    public List<Person> findAll(){
-        return personRepository.findAll();
+    public PersonDto findById(UUID id) {
+        return personRepository.findById(id)
+                .map(personMapper::toDto)
+                .orElse(null);
     }
-    public Person findById(int id){
-        Optional<Person> person = personRepository.findById(id);
-        return person.orElse(null);
-    }
+
     @Transactional
-    public void save(Person person){
+    public PersonDto save(PersonDto personDTO) {
+        PersonEntity person = personMapper.toEntity(personDTO);
         personRepository.save(person);
+        return personMapper.toDto(person);
     }
+
     @Transactional
-    public void update(int id, Person updatedPerson){
-        updatedPerson.setId(id);
+    public PersonDto update(UUID id, PersonDto updatedPersonDTO) {
+        PersonEntity existingPerson = personRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Person not found"));
+
+        PersonEntity updatedPerson = personMapper.toEntity(updatedPersonDTO);
+        updatedPerson.setId(existingPerson.getId());
         personRepository.save(updatedPerson);
+        return personMapper.toDto(updatedPerson);
     }
+
     @Transactional
-    public void delete(int id){
+    public PersonDto delete(UUID id) {
+        if (!personRepository.existsById(id)) {
+            throw new EntityNotFoundException("Person not found");
+        }
         personRepository.deleteById(id);
+        return personMapper.toDto(personRepository.findById(id).get());
     }
 
-    public List<Book> getBooksByPersonId(int id) {
-        return bookRepository.findBooksByPersonId(id);
+    public List<BookDto> getBooksByPersonId(UUID id) {
+        List<BookEntity> books = bookRepository.findBooksByPersonId(id);
+        return books.stream().map(bookMapper::toDto).toList();
     }
-
-
 }
+
